separate app components into pages / components
