+++ frontend/src/app/app.component.html
@@ -1,1 +1,1 @@
-<app-playfield></app-playfield>
+<router-outlet></router-outlet>
