+++ frontend/src/app/app.module.ts
@@ -3,14 +3,15 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { CardComponent } from './card/card.component';
-import { PlayfieldComponent } from './playfield/playfield.component';
-import { PlayerOverviewComponent } from './player-overview/player-overview.component';
-import { PlayersCardsComponent } from './players-cards/players-cards.component';
-import { CardFieldComponent } from './card-field/card-field.component';
-import { TurnIndicatorComponent } from './turn-indicator/turn-indicator.component';
-import { EmptyCardSpaceComponent } from './empty-card-space/empty-card-space.component';
-import { EmptyCardSpacesComponent } from './empty-card-spaces/empty-card-spaces.component';
+import { CardComponent } from './components/card/card.component';
+import { PlayfieldComponent } from './components/playfield/playfield.component';
+import { PlayerOverviewComponent } from './pages/player-overview/player-overview.component';
+import { PlayersCardsComponent } from './components/players-cards/players-cards.component';
+import { CardFieldComponent } from './components/card-field/card-field.component';
+import { TurnIndicatorComponent } from './components/turn-indicator/turn-indicator.component';
+import { EmptyCardSpaceComponent } from './components/empty-card-space/empty-card-space.component';
+import { EmptyCardSpacesComponent } from './components/empty-card-spaces/empty-card-spaces.component';
+import { UsernameSelectionComponent } from './components/username-selection/username-selection.component';
 
 @NgModule({
   declarations: [
