+++ frontend/src/app/components/card/card.component.html
