+++ frontend/src/app/components/card/card.component.scss
@@ -1,5 +1,5 @@
-@import './../../colors';
-@import './../../fonts';
+@import '../../../colors';
+@import '../../../fonts';
 
 .card {
   width: 8rem;
