+++ frontend/src/app/components/empty-card-space/empty-card-space.component.ts
