+++ frontend/src/app/components/empty-card-spaces/empty-card-spaces.component.scss
@@ -1,5 +1,5 @@
-@import './../../colors';
-@import './../../fonts';
+@import '../../../colors';
+@import '../../../fonts';
 
 :host {
   display: grid;
