+++ frontend/src/app/components/empty-card-spaces/empty-card-spaces.component.ts
