+++ frontend/src/app/components/players-cards/players-cards.component.html
