+++ frontend/src/app/components/playfield/playfield.component.scss
@@ -1,5 +1,5 @@
-@import './../../colors';
-@import './../../fonts';
+@import '../../../colors';
+@import '../../../fonts';
 
 :host {
   display: grid;
