+++ frontend/src/app/components/playfield/playfield.component.spec.ts
