+++ frontend/src/app/components/turn-indicator/turn-indicator.component.scss
@@ -1,5 +1,5 @@
-@import './../../colors';
-@import './../../fonts';
+@import '../../../colors';
+@import '../../../fonts';
 
 .display-turn {
   color: $tftf-orange;
