+++ frontend/src/app/components/turn-indicator/turn-indicator.component.spec.ts
