+++ frontend/src/app/components/turn-indicator/turn-indicator.component.ts
